Correct search images for several "all these tags"
When filter second radio is checked "Photos must match at least one filter" (OR) and a pair of "All these tags" (AND) is set. The first with tag1a AND tag1b, and the second with tag2a AND tag2b. I expect the SQL WHERE to be somehow like (tag1a AND tag1b) OR (tag2a AND tag2b).